requirements.txt
Flask
Flask-Bcrypt
PyJWT